Tutorial 4 - V1

+++ src/actions/CtrSteal.js
@@ -15,18 +15,21 @@ class CtrSteal extends Action {
 
         this.sprite.setVelocity(0, -40);
 
-        this.prey.setPosition(this.sprite.x, this.sprite.getBottomCenter().y).setVelocity(0);
-        this.prey.setState(States.STOLEN);
-
-        if (this.prey.y < -32) {
-            this.prey.kill();
-            this.prey.destroy();
-            
-            this.sprite.setY(-16)
-            this.setComplete();
+        if (this.prey) {
+            this.prey.setPosition(this.sprite.x, this.sprite.getBottomCenter().y).setVelocity(0);
+            this.prey.setState(States.STOLEN);
+    
+            if (this.prey.y < -32) {
+                this.prey.kill();
+                this.prey.destroy();
+                
+                this.sprite.setY(-16)
+                this.setComplete();
+            }
+            else if (!this.prey.isState(States.STOLEN))
+                this.setComplete();
         }
-        else if (!this.prey.isState(States.STOLEN))
-            this.setComplete();
+        
     }
 }
 export default CtrSteal;
